cast size_t returning functions to int

+++ src/util/info/info_get.c
@@ -72,7 +72,7 @@ int MPI_Info_get(MPI_Info info, char *key, int valuelen, char *value,
 		mpi_errno = MPIR_Err_create_code( MPI_SUCCESS, MPIR_ERR_RECOVERABLE, FCNAME, __LINE__, MPI_ERR_INFO_KEY,
 						  "**infokeynull", 0 );
 	    }
-	    else if ((keylen = strlen(key)) > MPI_MAX_INFO_KEY) {
+	    else if ((keylen = (int)strlen(key)) > MPI_MAX_INFO_KEY) {
 		mpi_errno = MPIR_Err_create_code( MPI_SUCCESS, MPIR_ERR_RECOVERABLE, FCNAME, __LINE__, MPI_ERR_INFO_KEY,
 						  "**infokeylong", 0 );
 	    } else if (keylen == 0) {
